Rename models/ to schemas/
